\name{rJPSGCS-package}
\alias{rJPSGCS-package}
\alias{rJPSGCS}
\docType{package}
\title{
R wrappers for JPSGCS gene drop programs.
}
\description{
R wrappers for gene drop programs in Alun Thomas' Java Programs for 
Statistical Genetics and Computational Statistics (JPSGCS). 
}
\details{
R wrappers for programs from JPSGCS required to 
simulate data by gene drop (\code{\link{GeneDrops}}) based on a 
fitted LD model (\code{\link{FitGMLD}}).
JPSGCS java source code was downloaded from Alun Thomas' website
(no longer active)
in March 2011. See the package NEWS file 
(\code{RShowDoc("NEWS", package="rJPSGCS")}) for a list of modifications 
needed to interface with R.
}
\author{
Sigal Blay, Jinko Graham, Brad McNeney and Annick Nembot-Simo

Maintainer: Brad McNeney <mcneney@sfu.ca>
}
\keyword{ package }
